Addition of Model Save/Model & Clearing ipynb outputs
Implemented the Save and Load model methods in the base gnn model class & the necessary _get_metadata() method for tracking the model architecture saved. Added the filtering in the GNNdataset for when a fitted encoder is passed to the dataset. Removed the outputs of the cells in the ipynb. Saved 3 trained models